backend/requirements.txt

coreapi==2.3.3
coreschema==0.0.4

#Django==5.0.3
Django==4.1.5
django-cors-headers==4.3.1
djangorestframework==3.14.0

requests==2.31.0
setuptools==69.2.0

virtualenv==20.25.1
